requirements.txt
opencv-python-headless>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0
imagehash>=4.3.1
streamlit>=1.29.0
imageio>=2.9.0
imageio-ffmpeg>=0.4.9
requests>=2.31.0
